[nasa/nos3#378] Created a fsw/shared directory and initial template for standalone;
