+++ .gitignore
@@ -1,0 +1,17 @@
+#
+# Metadata
+#
+.vagrant
+.vscode
+
+#
+# Build Files
+#
+build
+fprime/build
+fsw/standalone/build
+tmp
+core.*
+.cdskeyfile
+.reservedkeyfile
+.resetkeyfile
